Moved xml loaders to a better location
